Program to install HeliOS-specific updates
This program will install HeliOS-specific updates that are distributed in the form of ".deb" archives.  These archives are local configurations and programs to be installed on a HeliOS system.  They are (or have been traditionally) created with the awesome Effing Package Manager from Jordan Sissel (https://github.com/jordansissel/fpm).  They are tuned to the fixed directory structures of the HeliOS system, but are nothing more than standard ".deb" archives.  This program calls the standard "dpkg" utility to install them.  This utility expects to find them on an "APTonCD" repository volume mounted on the system.