HeliOS System Upgrade Program
This program is much like the "heliOS-updater" program.  Rather than installing HeliOS-specific ".deb" archives, it effects a full system upgrade from an "APTonCD" volume that it expects to find mounted on the system.  The purpose of this program -- like the "heliOS-updater" program -- is to enable system update maintenance of HeliOS systems that are deliberately isolated from public repositories on the Internet.